+++ vss_cli/plugins/request_plugins/change.py
@@ -98,6 +98,9 @@ def request_mgmt_change_retry(ctx: Configuration, rid):
         obj = ctx.retry_change_request(rid)
     columns = ctx.columns or const.COLUMNS_REQUEST_SUBMITTED
     ctx.echo(format_output(ctx, [obj], columns=columns, single=True))
+    # wait for request
+    if ctx.wait_for_requests:
+        ctx.wait_for_request_to(obj)
 
 
 @request_mgmt_change.group(