+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 # Changelog 📝
 
+## [v2021.5.0](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/tags/v2021.5.0) (2021-05-05)
+
+**New Features:**
+
+- [#418](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/418): ``compute vm set retire mk``: manage retirement requests for vms.
+- [#418](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/418): ``compute vm set retire confirm``: manage retirement requests for vms.
+- [#418](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/418): ``compute vm set retire cancel``: manage retirement requests for vms.
+- [#418](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/418): ``compute vm set retire send``: manage retirement requests for vms.
+- [#419](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/419): ``compute vm get retire``: get retirement requests for vm.
+- [#420](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/420): ``request retire ls``: list retirement requests.
+- [#420](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/420): ``request retire get``: get retirement request info.
+- [#420](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/420): ``request retire confirm``: confirm retirement request.
+- [#420](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/420): ``request retire cancel``: cancel retirement request.
+- [#420](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/420): ``request retire send``: send notification for a retirement request.
+
+**Improvements:**
+
+- [#417](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/417): ``core``: move from ``semver`` to ``calver``.
+- [#423](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/423): ``compute vm set custom-spec``: allow multiple ``--dns-suffix`` options for dns search settings.
+- [#424](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/424): ``compute vm mk shell``: create VM with retirement request ``--retire-type``, ``--retire-value``, ``--retire-warning``.
+- [#424](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/424): ``compute vm mk from-clone``: create VM with retirement request ``--retire-type``, ``--retire-value``, ``--retire-warning``.
+- [#424](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/424): ``compute vm mk from-spec``: create VM with retirement request ``--retire-type``, ``--retire-value``, ``--retire-warning``.
+- [#424](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/424): ``compute vm mk from-image``: create VM with retirement request ``--retire-type``, ``--retire-value``, ``--retire-warning``.
+- [#424](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/424): ``compute vm mk from-template``: create VM with retirement request ``--retire-type``, ``--retire-value``, ``--retire-warning``.
+
+**Bug Fixes:**
+
+- [#421](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/421): ``request new retry``: ignores ``--wait`` option.
+- [#422](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/422): ``request change retry``: ignores ``--wait`` option.
+
 ## [v0.12.1](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/tags/v0.12.1) (2021-04-15)
 
 **Improvements:**